requirements.txt
requests
python-dotenv
discord.py
bs4
lxml
requests
RedditReader
